APIs
This repository is a collection of three unique APIs using jQuery to create a basic search function for videos, maps and e-commerce books. Once code is downloaded you can use each of the three functions to find suitable / desired youtube video's, books or use the code to pin point an exact area of google maps you want to highlight. thanks